Refactors Testimonial and TestimonialForm
